refactor: promote shellQuoteJoin to plugin.ShellQuoteJoin
The quoting logic (single-quote each arg, join with spaces) was
duplicated between cmd/run.go and the inline block added in lifecycle.go.
Promoted to plugin.ShellQuoteJoin, which builds on the existing
plugin.ShellQuote. Both callers updated; cmd/run_test.go tests moved
to internal/plugin/shell_test.go alongside ShellQuote tests.
